Bug fix: clear listener for parseBtn
Previous code attched an eventlistener for parseBtn everytime the user selects a different dropdown menu, so that it remembers the history of the bookmark folders selected by the user and would append all urls from all the folders to a chosen Google Sheet. The new code ensures that all previous listeners are removed before creating a new one.